dose engine can now use single precision calculations

+++ matRad/rayTracing/matRad_siddonRayTracer.m
@@ -227,9 +227,9 @@ i_mm = sourcePoint(1) + alphas_mid*(targetPoint(1) - sourcePoint(1));
 j_mm = sourcePoint(2) + alphas_mid*(targetPoint(2) - sourcePoint(2));
 k_mm = sourcePoint(3) + alphas_mid*(targetPoint(3) - sourcePoint(3));
 % then convert to voxel index
-i = round(i_mm/resolution.x);
-j = round(j_mm/resolution.y);
-k = round(k_mm/resolution.z);
+i = round(double(i_mm/resolution.x));
+j = round(double(j_mm/resolution.y));
+k = round(double(k_mm/resolution.z));
 
 % Handle numerical instabilities at the borders.
 i(i<=0) = 1; j(j<=0) = 1; k(k<=0) = 1;
